% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{apply_mapshaper_commands}
\alias{apply_mapshaper_commands}
\title{#run_mapshaper_command
#
#}
\usage{
apply_mapshaper_commands(command, data)
}
\arguments{
\item{command}{#
#}

\item{data}{#}

\item{command}{}

\item{data}{}
}
\value{
geojson
#

geojson
}
\description{
#run_mapshaper_command
#
#
}

