% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clip_erase.R
\name{clip}
\alias{clip}
\title{clip}
\usage{
clip(target, clip)
}
\arguments{
\item{target}{the target layer. Can be json or sp class}

\item{clip}{the clipping layer. Can be json or SpatialPolygonsDataFrame}
}
\value{
clipped target in the same class as the input target
}
\description{
clip
}

