% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissolve.R
\name{dissolve}
\alias{dissolve}
\title{Aggregate shapes in a polygon layer.}
\usage{
dissolve(sp_obj, field = NULL, sum_fields = NULL, copy_fields = NULL,
  snap = TRUE)
}
\arguments{
\item{sp_obj}{spatial object to dissolve - can be one of the Spatial classes (e.g., SpatialPolygonsDataFrame) or class json}

\item{field}{}

\item{sum_fields}{}

\item{copy_fields}{}

\item{snap}{Snap together vertices within a small distance threshold to
fix small coordinate misalignment in adjacent polygons. Default
\code{TRUE}.}
}
\value{
a simplified representation of the geometry in the same class as the input
}
\description{
Aggregates using specified field, or all shapes if no field is given
}

