% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{drop_null_geometries}
\alias{drop_null_geometries}
\title{Drop features from a geo_list FeatureCollection with null geometries}
\usage{
drop_null_geometries(x)
}
\arguments{
\item{x}{a geo_list FeatureCollection}
}
\value{
a geo_list FeatureCollection with Features with null geometries
removed
}
\description{
Drop features from a geo_list FeatureCollection with null geometries
}

