% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clip_erase.R
\name{erase}
\alias{erase}
\title{erase}
\usage{
erase(target, erase)
}
\arguments{
\item{target}{the target layer. Can be json or sp class}

\item{erase}{the erase layer. Can be json or SpatialPolygonsDataFrame}
}
\value{
erased target in the same format as the input target
}
\description{
erase
}

