% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_erase.R
\name{ms_clip}
\alias{ms_clip}
\alias{ms_clip.SpatialLines}
\alias{ms_clip.SpatialPoints}
\alias{ms_clip.SpatialPolygons}
\alias{ms_clip.character}
\alias{ms_clip.geo_json}
\alias{ms_clip.geo_list}
\title{Remove features or portions of features that fall outside a clipping area.}
\usage{
ms_clip(target, clip = NULL, bbox = NULL, force_FC = TRUE)

\method{ms_clip}{character}(target, clip = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_clip}{geo_json}(target, clip = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_clip}{geo_list}(target, clip = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_clip}{SpatialPolygons}(target, clip = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_clip}{SpatialLines}(target, clip = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_clip}{SpatialPoints}(target, clip = NULL, bbox = NULL,
  force_FC = TRUE)
}
\arguments{
\item{target}{the target layer from which to remove portions. Can be \code{geo_json} or \code{sp} class}

\item{clip}{the clipping layer. Can be \code{geo_json} or \code{SpatialPolygons*}}

\item{bbox}{supply a bounding box instead of a clippling layer to extract from
the target layer. Supply as a numeric vector: \code{c(minX, minY, maxX, maxY)}.}

\item{force_FC}{should the output be forced to be a \code{FeatureCollection} even
if there are no attributes? Default \code{TRUE}.
 \code{FeatureCollections} are more compatible with \code{rgdal::readOGR} and
 \code{geojsonio::geojson_sp}. If \code{FALSE} and there are no attributes associated with
 the geometries, a \code{GeometryCollection} will be output. Ignored for \code{Spatial}
 objects, as the output is always the same as the input.}
}
\value{
clipped target in the same class as the input target
}
\description{
Removes portions of the target layer that fall outside the clipping layer or bounding box.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: For character representations of geojson (for example
if you used \code{readLines} to read in a geojson file)

\item \code{geo_json}: Method for geo_json objects

\item \code{geo_list}: Method for geo_list objects

\item \code{SpatialPolygons}: Method for SpatialPolygons* objects

\item \code{SpatialLines}: Method for SpatialLines* objects

\item \code{SpatialPoints}: Method for SpatialPoints* objects
}}

