% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{ms_filter}
\alias{ms_filter}
\alias{ms_filter.SpatialLinesDataFrame}
\alias{ms_filter.SpatialPointsDataFrame}
\alias{ms_filter.SpatialPolygonsDataFrame}
\alias{ms_filter.character}
\alias{ms_filter.geo_json}
\alias{ms_filter.geo_list}
\title{Filter features based on attributes}
\usage{
ms_filter(input, filter = NULL, remove_null_geometries = TRUE,
  force_FC = TRUE)

\method{ms_filter}{character}(input, filter = NULL,
  remove_null_geometries = TRUE, force_FC = TRUE)

\method{ms_filter}{geo_json}(input, filter = NULL,
  remove_null_geometries = TRUE, force_FC = TRUE)

\method{ms_filter}{geo_list}(input, filter = NULL,
  remove_null_geometries = TRUE, force_FC = TRUE)

\method{ms_filter}{SpatialPolygonsDataFrame}(input, filter = NULL,
  remove_null_geometries = TRUE, force_FC)

\method{ms_filter}{SpatialLinesDataFrame}(input, filter = NULL,
  remove_null_geometries = TRUE, force_FC)

\method{ms_filter}{SpatialPointsDataFrame}(input, filter = NULL,
  remove_null_geometries = TRUE, force_FC)
}
\arguments{
\item{input}{spatial object to filter - can be one of the \code{Spatial}
classes (e.g., \code{SpatialPolygonsDataFrame}) or class \code{geo_json}}

\item{filter}{expression to apply. Can be a character vector of individual
expressions, which will be combined with logical \code{AND (&)}, or a
single string combing expressions with \code{&} and/or \code{|}.}

\item{force_FC}{should the output be forced to be a \code{FeatureCollection}
even if there are no attributes? Default \code{TRUE}.
\code{FeatureCollections} are more compatible with \code{rgdal::readOGR}
and \code{geojsonio::geojson_sp}. If \code{FALSE} and there are no
attributes associated with the geometries, a \code{GeometryCollection} will
be output. Ignored for \code{Spatial} objects, as a
\code{Spatial*DataFrame} is always the output.}
}
\value{
object with only specified features retained, in the same class as the input
}
\description{
Apply a boolean expression to the attributes of each feature, removeing those
which evaluate to \code{FALSE}
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: For character representations of geojson (for example
if you used \code{readLines} to read in a geojson file)

\item \code{geo_json}: Method for geo_json

\item \code{geo_list}: Method for geo_list

\item \code{SpatialPolygonsDataFrame}: Method for SpatialPolygonsDataFrame

\item \code{SpatialLinesDataFrame}: Method for SpatialLinesDataFrame

\item \code{SpatialPointsDataFrame}: Method for SpatialPointsDataFrame
}}

