% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fields.R
\name{ms_filter_fields}
\alias{ms_filter_fields}
\alias{ms_filter_fields.SpatialLinesDataFrame}
\alias{ms_filter_fields.SpatialPointsDataFrame}
\alias{ms_filter_fields.SpatialPolygonsDataFrame}
\alias{ms_filter_fields.character}
\alias{ms_filter_fields.geo_json}
\alias{ms_filter_fields.geo_list}
\title{Delete fields in the attribute table}
\usage{
ms_filter_fields(input, fields)

\method{ms_filter_fields}{character}(input, fields)

\method{ms_filter_fields}{geo_json}(input, fields)

\method{ms_filter_fields}{geo_list}(input, fields)

\method{ms_filter_fields}{SpatialPolygonsDataFrame}(input, fields)

\method{ms_filter_fields}{SpatialPointsDataFrame}(input, fields)

\method{ms_filter_fields}{SpatialLinesDataFrame}(input, fields)
}
\arguments{
\item{input}{spatial object to filter - can be a \code{Spatial*DataFrame} or
class \code{geo_json} or \code{geo_list}}

\item{fields}{character vector of fields to retain.}
}
\value{
object with only specified attributes retained, in the same class as
  the input
}
\description{
Removes all fields except those listed in the \code{fields} parameter
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: For character representations of geojson (for example
if you used \code{readLines} to read in a geojson file)

\item \code{geo_json}: Method for geo_json

\item \code{geo_list}: Method for geo_list

\item \code{SpatialPolygonsDataFrame}: Method for SpatialPolygonsDataFrame

\item \code{SpatialPointsDataFrame}: Method for SpatialPointsDataFrame

\item \code{SpatialLinesDataFrame}: Method for SpatialLinesDataFrame
}}

