% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{ms_lines}
\alias{ms_lines}
\title{Convert polygons to topological boundaries (lines)}
\usage{
ms_lines(input, fields = NULL, force_FC = TRUE)
}
\arguments{
\item{input}{input polygons object to convert to inner lines. One of:
\itemize{
 \item \code{geo_json} or \code{character} polygons;
 \item \code{geo_list} polygons;
 \item \code{SpatialPolygons*}
 }}

\item{fields}{character vector of field names. If left as \code{NULL}
(default), external (unshared) boundaries are attributed as TYPE 0 and
internal (shared) boundaries are TYPE 1. Giving a field name adds an
intermediate level of hierarchy at TYPE 1, with the lowest-level internal
boundaries set to TYPE 2. Supplying a character vector of field names adds
additional levels of hierarchy.}

\item{force_FC}{should the output be forced to be a \code{FeatureCollection}
even if there are no attributes? Default \code{TRUE}.
\code{FeatureCollections} are more compatible with \code{rgdal::readOGR}
and \code{geojsonio::geojson_sp}. If \code{FALSE} and there are no
attributes associated with the geometries, a \code{GeometryCollection} will
be output. Ignored for \code{Spatial} objects.}
}
\value{
topological boundaries as lines, in the same class as the input
}
\description{
Convert polygons to topological boundaries (lines)
}

