% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapshaper-package.R
\docType{package}
\name{rmapshaper-package}
\alias{rmapshaper}
\alias{rmapshaper-package}
\title{rmapshaper: Client for 'mapshaper' for 'Geospatial' Operations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Edit and simplify 'geojson', 'Spatial', and 'sf' objects. This is wrapper around the 'mapshaper' 'JavaScript' library by Matthew Bloch \url{https://github.com/mbloch/mapshaper/} to perform topologically-aware polygon simplification, as well as other operations such as clipping, erasing, dissolving, and converting 'multi-part' to 'single-part' geometries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ateucher/rmapshaper}
  \item \url{http://andyteucher.ca/rmapshaper/}
  \item Report bugs at \url{https://github.com/ateucher/rmapshaper/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})

Authors:
\itemize{
  \item Kenton Russell \email{kent.russell@timelyportfolio.com} (JavaScript support)
}

Other contributors:
\itemize{
  \item Matthew Bloch (mapshaper Javascript library) [copyright holder]
}

}
\keyword{internal}
