% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapshaper.R
\docType{package}
\name{rmapshaper}
\alias{rmapshaper}
\alias{rmapshaper-package}
\title{rmapshaper: A package for editing spatial objects.}
\description{
Topologically-aware polygon simplification.
}
\section{rmapshaper functions}{


ms_simplify

ms_explode

ms_dissolve

ms_clip

ms_erase
}

