% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{run_mapshaper_command}
\alias{run_mapshaper_command}
\title{run_mapshaper_command}
\usage{
run_mapshaper_command(data, command)
}
\arguments{
\item{data}{}

\item{command}{}
}
\value{
geojson
}
\description{
run_mapshaper_command
}

