% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{apply_mapshaper_commands}
\alias{apply_mapshaper_commands}
\title{Apply a mapshaper command string to a geojson object}
\usage{
apply_mapshaper_commands(command, data)
}
\arguments{
\item{command}{command string}

\item{data}{geojson object}
}
\value{
geojson
}
\description{
Apply a mapshaper command string to a geojson object
}

