% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{apply_mapshaper_commands}
\alias{apply_mapshaper_commands}
\title{apply_mapshaper_commands}
\usage{
apply_mapshaper_commands(command, data)
}
\arguments{
\item{command}{command string}

\item{data}{geojson object}
}
\value{
geojson
}
\description{
apply_mapshaper_commands
}

