% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sys_mapshaper}
\alias{check_sys_mapshaper}
\title{Check the system mapshaper}
\usage{
check_sys_mapshaper(verbose = TRUE)
}
\arguments{
\item{verbose}{Print a message stating mapshaper's current version? Default \code{TRUE}}
}
\value{
TRUE (with a message) if appropriate version is installed, otherwise throws an error
}
\description{
Check the system mapshaper
}
