% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sys_mapshaper}
\alias{check_sys_mapshaper}
\title{Check the system mapshaper}
\usage{
check_sys_mapshaper(command = "mapshaper-xl", verbose = TRUE)
}
\arguments{
\item{command}{either "mapshaper-xl" (default) or "mapshaper"}

\item{verbose}{Print a message stating mapshaper's current version? Default \code{TRUE}}
}
\value{
character path to mapshaper executable if appropriate version is installed, otherwise throws an error
}
\description{
Check the system mapshaper
}
