% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_null_geometries.R
\name{drop_null_geometries}
\alias{drop_null_geometries}
\title{Drop features from a \code{geo_list} or \code{geo_json} FeatureCollection with null geometries}
\usage{
drop_null_geometries(x)
}
\arguments{
\item{x}{a \code{geo_list} or \code{geo_json} FeatureCollection}
}
\value{
a \code{geo_list} or \code{geo_json} FeatureCollection with Features with null geometries
removed
}
\description{
Drop features from a \code{geo_list} or \code{geo_json} FeatureCollection with null geometries
}
