% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_erase.R
\name{ms_clip}
\alias{ms_clip}
\alias{ms_clip.SpatialPolygonsDataFrame}
\alias{ms_clip.geo_json}
\alias{ms_clip.geo_list}
\title{Remove features or portions of features that fall outside a clipping area.}
\usage{
ms_clip(target, clip = NULL, bbox = NULL, force_FC = TRUE)

\method{ms_clip}{geo_json}(target, clip = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_clip}{geo_list}(target, clip = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_clip}{SpatialPolygonsDataFrame}(target, clip = NULL, bbox = NULL,
  force_FC = TRUE)
}
\arguments{
\item{target}{the target layer from which to remove portions. Can be geo_json or sp class}

\item{clip}{the clipping layer. Can be geo_json or SpatialPolygonsDataFrame}

\item{bbox}{supply a bounding box instead of a clippling layer to extract from
the target layer. Supply as a numeric vector: \code{c(minX, minY, maxX, maxY)}.}

\item{force_FC}{should the output be forced to be a FeatureCollection (or
Spatial*DataFrame) even if there are no attributes? Default \code{TRUE}.
FeatureCollections are more compatible with rgdal::readOGR and
geojsonio::geojson_sp. If FALSE and there are no attributes associated with
the geometries, a GeometryCollection (or Spatial object with no dataframe)
will be output.}
}
\value{
clipped target in the same class as the input target
}
\description{
Removes portions of the target layer that fall outside the clipping layer or bounding box.
}
\section{Methods (by class)}{
\itemize{
\item \code{geo_json}: Method for geo_json objects

\item \code{geo_list}: Method for geo_list objects

\item \code{SpatialPolygonsDataFrame}: Method for SpatialPolygonsDataFrame objects
}}

