% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve.R
\name{ms_dissolve}
\alias{ms_dissolve}
\alias{ms_dissolve.SpatialPolygonsDataFrame}
\alias{ms_dissolve.character}
\alias{ms_dissolve.geo_json}
\alias{ms_dissolve.geo_list}
\title{Aggregate shapes in a polygon layer.}
\usage{
ms_dissolve(input, field = NULL, sum_fields = NULL, copy_fields = NULL,
  snap = TRUE, force_FC = TRUE)

\method{ms_dissolve}{character}(input, field = NULL, sum_fields = NULL,
  copy_fields = NULL, snap = TRUE, force_FC = TRUE)

\method{ms_dissolve}{geo_json}(input, field = NULL, sum_fields = NULL,
  copy_fields = NULL, snap = TRUE, force_FC = TRUE)

\method{ms_dissolve}{geo_list}(input, field = NULL, sum_fields = NULL,
  copy_fields = NULL, snap = TRUE, force_FC = TRUE)

\method{ms_dissolve}{SpatialPolygonsDataFrame}(input, field = NULL,
  sum_fields = NULL, copy_fields = NULL, snap = TRUE, force_FC = TRUE)
}
\arguments{
\item{input}{spatial object to dissolve - can be an be a
\code{SpatialPolygonsDataFrame} or class \code{geo_json} or
\code{geo_list}}

\item{field}{the field to dissolve on}

\item{sum_fields}{fields to sum}

\item{copy_fields}{fields to copy. The first instance of each field will be
copied to the aggregated feature.}

\item{snap}{Snap together vertices within a small distance threshold to fix
small coordinate misalignment in adjacent polygons. Default \code{TRUE}.}

\item{force_FC}{should the output be forced to be a \code{FeatureCollection} even
if there are no attributes? Default \code{TRUE}.
 \code{FeatureCollections} are more compatible with \code{rgdal::readOGR} and
 \code{geojsonio::geojson_sp}. If \code{FALSE} and there are no attributes associated with
 the geometries, a \code{GeometryCollection} will be output. Ignored for \code{Spatial}
 objects, as a \code{Spatial*DataFrame} is always the output.}
}
\value{
the same class as the input
}
\description{
Aggregates using specified field, or all shapes if no field is given
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: For character representations of geojson (for example
if you used \code{readLines} to read in a geojson file)

\item \code{geo_json}: For geo_json objects

\item \code{geo_list}: For geo_list objects

\item \code{SpatialPolygonsDataFrame}: For SpatialPolygonsDataFrame objects
}}

