% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve.R
\name{ms_dissolve}
\alias{ms_dissolve}
\title{Aggregate shapes in a polygon layer.}
\usage{
ms_dissolve(input, field = NULL, sum_fields = NULL, copy_fields = NULL,
  snap = TRUE)
}
\arguments{
\item{input}{spatial object to dissolve - can be one of the Spatial classes (e.g., SpatialPolygonsDataFrame) or class json}

\item{field}{the field to dissolve on}

\item{sum_fields}{fields to sum}

\item{copy_fields}{fields to copy. The first instance of each field will be copied to the aggregated feature.}

\item{snap}{Snap together vertices within a small distance threshold to
fix small coordinate misalignment in adjacent polygons. Default
\code{TRUE}.}
}
\value{
the same class as the input
}
\description{
Aggregates using specified field, or all shapes if no field is given
}

