% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_erase.R
\name{ms_erase}
\alias{ms_erase}
\title{erase}
\usage{
ms_erase(target, erase)
}
\arguments{
\item{target}{the target layer. Can be json or sp class}

\item{erase}{the erase layer. Can be json or SpatialPolygonsDataFrame}
}
\value{
erased target in the same format as the input target
}
\description{
erase
}

