% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_erase.R
\name{ms_erase}
\alias{ms_erase}
\alias{ms_erase.SpatialLinesDataFrame}
\alias{ms_erase.SpatialPointsDataFrame}
\alias{ms_erase.SpatialPolygonsDataFrame}
\alias{ms_erase.character}
\alias{ms_erase.geo_json}
\alias{ms_erase.geo_list}
\title{Remove features or portions of features that fall inside a specified area}
\usage{
ms_erase(target, erase = NULL, bbox = NULL, force_FC = TRUE)

\method{ms_erase}{character}(target, erase = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_erase}{geo_json}(target, erase = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_erase}{geo_list}(target, erase = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_erase}{SpatialPolygonsDataFrame}(target, erase = NULL,
  bbox = NULL, force_FC = TRUE)

\method{ms_erase}{SpatialLinesDataFrame}(target, erase = NULL, bbox = NULL,
  force_FC = TRUE)

\method{ms_erase}{SpatialPointsDataFrame}(target, erase = NULL, bbox = NULL,
  force_FC = TRUE)
}
\arguments{
\item{target}{the target layer from which to remove portions. Can be \code{geo_json} or \code{sp} class}

\item{erase}{the erasing layer. Can be \code{geo_json} or \code{SpatialPolygonsDataFrame}.}

\item{bbox}{supply a bounding box instead of an erasing layer to remove from
the target layer. Supply as a numeric vector: \code{c(minX, minY, maxX, maxY)}.}

\item{force_FC}{should the output be forced to be a \code{FeatureCollection} even
if there are no attributes? Default \code{TRUE}.
 \code{FeatureCollections} are more compatible with \code{rgdal::readOGR} and
 \code{geojsonio::geojson_sp}. If \code{FALSE} and there are no attributes associated with
 the geometries, a \code{GeometryCollection} will be output. Ignored for \code{Spatial}
 objects, as a \code{Spatial*DataFrame} is always the output.}
}
\value{
erased target in the same format as the input target
}
\description{
Removes portions of the target layer that fall inside the erasing layer or bounding box.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: For character representations of geojson (for example
if you used \code{readLines} to read in a geojson file)

\item \code{geo_json}: Method for geo_json objects

\item \code{geo_list}: Method for geo_list objects

\item \code{SpatialPolygonsDataFrame}: Method for SpatialPolygonsDataFrame objects

\item \code{SpatialLinesDataFrame}: Method for SpatialLinesDataFrame objects

\item \code{SpatialPointsDataFrame}: Method for SpatialPointsDataFrame objects
}}

