% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_erase.R
\name{ms_erase}
\alias{ms_erase}
\title{erase}
\usage{
ms_erase(target, erase, force_FC = TRUE)
}
\arguments{
\item{target}{the target layer. Can be geo_json or sp class}

\item{erase}{the erase layer. Can be geo_json or SpatialPolygonsDataFrame.}

\item{force_FC}{should the output be forced to be a FeatureCollection (or
Spatial*DataFrame) even if there are no attributes? Default \code{TRUE}.
FeatureCollections are more compatible with rgdal::readOGR and
geojsonio::geojson_sp. If FALSE and there are no attributes associated with
the geometries, a GeometryCollection (or Spatial object with no dataframe)
will be output.}
}
\value{
erased target in the same format as the input target
}
\description{
erase
}

