% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explode.R
\name{ms_explode}
\alias{ms_explode}
\title{Convert multipart polygons to singlepart}
\usage{
ms_explode(x)
}
\arguments{
\item{x}{}
}
\value{
json or SpatialPolygonsDataFrame
}
\description{
Convert multipart polygons to singlepart
}

