% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{ms_explode}
\alias{ms_explode}
\alias{ms_explode.SpatialPolygonsDataFrame}
\alias{ms_explode.geo_json}
\alias{ms_explode.geo_list}
\title{Convert multipart polygons to singlepart}
\usage{
ms_explode(input, force_FC = TRUE)

\method{ms_explode}{geo_json}(input, force_FC = TRUE)

\method{ms_explode}{geo_list}(input, force_FC = TRUE)

\method{ms_explode}{SpatialPolygonsDataFrame}(input, force_FC)
}
\arguments{
\item{input}{\code{geojson} or \code{SpatialPolygonsDataFrame} object containing
multipart polygons}

\item{force_FC}{should the output be forced to be a \code{FeatureCollection} even
if there are no attributes? Default \code{TRUE}.
 \code{FeatureCollections} are more compatible with \code{rgdal::readOGR} and
 \code{geojsonio::geojson_sp}. If \code{FALSE} and there are no attributes associated with
 the geometries, a \code{GeometryCollection} will be output. Ignored for \code{Spatial}
 objects, as a \code{Spatial*DataFrame} is always the output.}
}
\value{
same class as input
}
\description{
For objects of class \code{Spatial} (e.g., \code{SpatialPolygonsDataFrame}), this is
simply a wrapper around \code{sp::disaggregate}.
}
\section{Methods (by class)}{
\itemize{
\item \code{geo_json}: Method for geo_json

\item \code{geo_list}: Method for geo_list

\item \code{SpatialPolygonsDataFrame}: Method for SpatialPolygonsDataFrame
}}

