% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fields.R
\name{ms_filter_fields}
\alias{ms_filter_fields}
\title{Delete fields in the attribute table}
\usage{
ms_filter_fields(input, fields)
}
\arguments{
\item{input}{spatial object to filter fields on. One of:
\itemize{
 \item \code{geo_json} or \code{character} points, lines, or polygons;
 \item \code{geo_list} points, lines, or polygons;
 \item \code{SpatialPolygonsDataFrame}, \code{SpatialLinesDataFrame}, \code{SpatialPointsDataFrame}
 }}

\item{fields}{character vector of fields to retain.}
}
\value{
object with only specified attributes retained, in the same class as
  the input
}
\description{
Removes all fields except those listed in the \code{fields} parameter
}

