% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_lines.R
\name{ms_innerlines}
\alias{ms_innerlines}
\alias{ms_innerlines.SpatialPolygonsDataFrame}
\alias{ms_innerlines.character}
\alias{ms_innerlines.geo_json}
\alias{ms_innerlines.geo_list}
\title{Create a line layer consisting of shared boundaries with no attribute data}
\usage{
ms_innerlines(input, force_FC = TRUE)

\method{ms_innerlines}{character}(input, force_FC = TRUE)

\method{ms_innerlines}{geo_json}(input, force_FC = TRUE)

\method{ms_innerlines}{geo_list}(input, force_FC = TRUE)

\method{ms_innerlines}{SpatialPolygonsDataFrame}(input, force_FC)
}
\arguments{
\item{input}{input polygons object to dissolve - can be a
\code{SpatialPolygonsDataFrame} or class \code{geo_json} or
\code{geo_list}}

\item{force_FC}{should the output be forced to be a \code{FeatureCollection}
even if there are no attributes? Default \code{TRUE}.
\code{FeatureCollections} are more compatible with \code{rgdal::readOGR}
and \code{geojsonio::geojson_sp}. If \code{FALSE} and there are no
attributes associated with the geometries, a \code{GeometryCollection} will
be output. Ignored for \code{Spatial} objects, as a
\code{SpatialLinesDataFrame} is always the output.}
}
\value{
lines in the same class as the input layer
}
\description{
Create a line layer consisting of shared boundaries with no attribute data
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: For character representations of geojson (for example
if you used \code{readLines} to read in a geojson file)

\item \code{geo_json}: Method for geo_json

\item \code{geo_list}: Method for geo_list

\item \code{SpatialPolygonsDataFrame}: Method for SpatialPolygonsDataFrame
}}

