% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{ms_lines}
\alias{ms_lines}
\alias{ms_lines.SpatialPolygons}
\alias{ms_lines.character}
\alias{ms_lines.geo_json}
\alias{ms_lines.geo_list}
\title{Convert polygons to topological boundaries (lines)}
\usage{
ms_lines(input, fields = NULL, force_FC = TRUE)

\method{ms_lines}{character}(input, fields = NULL, force_FC = TRUE)

\method{ms_lines}{geo_json}(input, fields = NULL, force_FC = TRUE)

\method{ms_lines}{geo_list}(input, fields = NULL, force_FC = TRUE)

\method{ms_lines}{SpatialPolygons}(input, fields = NULL, force_FC)
}
\arguments{
\item{input}{input polygons object to convert to lines - can be a
\code{SpatialPolygons*} or class \code{geo_json} or
\code{geo_list}}

\item{fields}{character vector of field names. If left as \code{NULL}
(default), external (unshared) boundaries are attributed as TYPE 0 and
internal (shared) boundaries are TYPE 1. Giving a field name adds an
intermediate level of hierarchy at TYPE 1, with the lowest-level internal
boundaries set to TYPE 2. Supplying a character vector of field names adds
additional levels of hierarchy.}

\item{force_FC}{should the output be forced to be a \code{FeatureCollection}
even if there are no attributes? Default \code{TRUE}.
\code{FeatureCollections} are more compatible with \code{rgdal::readOGR}
and \code{geojsonio::geojson_sp}. If \code{FALSE} and there are no
attributes associated with the geometries, a \code{GeometryCollection} will
be output. Ignored for \code{Spatial} objects.}
}
\value{
topological boundaries as lines, in the same class as the input
}
\description{
Convert polygons to topological boundaries (lines)
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: For character representations of geojson (for example
if you used \code{readLines} to read in a geojson file)

\item \code{geo_json}: Method for geo_json

\item \code{geo_list}: Method for geo_list

\item \code{SpatialPolygons}: Method for SpatialPolygons
}}

