% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Topologically-aware simplification of spatial objects.}
\usage{
simplify(sp_obj, keep = 0.05, method = "vis", keep_shapes = TRUE,
  no_repair = FALSE, auto_snap = TRUE)
}
\arguments{
\item{sp_obj}{spatial object to simplify}

\item{keep}{proportion of points to retain (0-1; default 0.05)}

\item{method}{simplification method to use: \code{"vis"} for Visvalingam
algorithm (default), or \code{"dp"} for Douglas-Peuker algorithm. See this
\url{https://github.com/mbloch/mapshaper/wiki/Simplification-Tips}{link}
for more information.}

\item{keep_shapes}{Prevent polygon features from disappearing at high
simplification (default \code{TRUE})}

\item{no_repair}{disable intersection repair after simplification (default
\code{FALSE}).}

\item{auto_snap}{Snap together vertices within a small distance threshold to
  fix small coordinate misalignment in adjacent polygons. Default
  \code{TRUE}.}
}
\value{
an \code{sp} object
}
\description{
Uses \href{https://github.com/mbloch/mapshaper}{mapshaper} to simplify
polygons. It is a Node library, so you need to have Node installed to use it:
\url{https://nodejs.org/download/}. Then install mapshaper on the command
line with: \code{npm install -g mapshaper}.
}

